Complete test_scoring module
Added testing for fail cases where dice submitted for scoring were in-
valid for the category selected.

Sample dice and categories and expected values are in a dict with lists.
Testing of all categories is iterated through by the test function.

pytest was utilized in the testing of the test_scoring module.