Add check for verifying player selection has not already been used
In scorepad object add attribute the stores all the valid choices into
a list. As the game progresses a player's submitted choice is checked to
see if it is part of the accepted list of choices as well as confirm the
selection has not already been used in previous turns.

If choice is valid and after score selection has been updated in the
scorepad object, delete the selection from the available choices attri-
bute.