Fix free roll issue with five of a kind bonus category
When player selects a category, the five of a kind bonus
category allows player to select it regardless of player
having a five of a kind score and effectively finish the
turn without scoring or zeroing a score if there is at least
one score category available. This update checks to see if
a five of a kind has been scored and that it is not zero
before appending it to the available score category menu
list.