Webdev - Merge latest web app code (#3)
* Add code to render roll one results and then pass info to roll 2

After displaying initial roll of five dice need to request dice to be
held and include in next roll (#2). Need to pass object with game data
and track turn and dice info to roll two html and route.

* Add ability to pass object from one view to another

Need to determine best method to pass the scorepad from view to view
in order to keep the scoring and tracking information for the game and
player.

* Move routes from app.py to webgame module and bluprint

Restructure web organization to utilize Blueprints. Still need to deter-
mine best way to keep scorepad object data when redirecting to different
templates throughout a turn and during the whole game. May need to
utilize sqlite since it can be persistent during session.

* Finish out build of the first roll game process and HTML rendering

The webgame subpackage will process the player turn for rolling three
times. Roll one rolls all five dice and then player selects which five
to keep. Not selecting any dice will re-roll all five. After the form is
submitted, roll two route will render the result of the second roll.

Second and third roll processing are pending.

Scoring section is next.

This commit adds the following to the dice subpackage:

- roll_five_dice() - Roll all five dice. This occurs at the beginning of
  every turn.

- dice_png_list(dice_list) - Generate a list of the image file names
  based on the dice list to render.

* Finish initial build of second roll processing and rendering

Dice kept from first roll are included in second roll. After submit
action, second roll is rendered and process and forms to select dice
to keep is repeated.

* Consolidate the rolls per turn into one route

All rolls will now go through one route. When player reaches three
rolls the game play will need to continue to a score display and form to
select which score category to assign the dice roll.

* Start build of scoring section

After third roll (or directly from first or second roll once buttion is
in place) render the category scores and totals. Player will then select
a category to assign the current dice. This will kick of the scorepad
update process.

* Build out main game flow to end of game

Complete overall functionality to go from first turn to susequent turns
by:
   - processing a turn of up to three rolls
   - displaying final dice of turn, display current categories score
   - Dropdown list of available score categories from which to select
   - Upon selection, take choice and update the scorepad object with
     score and update tracking flag for the corresponding score.
   - Evaluate remaining available scores in scorepad
         - If categories remain reset turn tracking and dice list
           and repeat the turn process
         - If player has updated all score categories, end the game
           and display final score.

* Enable Keep All Dice button

Keep All Dice and Select Score button will allow a player to keep
current dice after roll one or two and skip to the category score selec-
tion stage.

* Re-organize score display and consolidate scoredisplay page

Create a smaller footprint for the score display. Utilize a single
score display HTML and use routes to set flag so that appropriate
headers and buttons are rendered.

Also, fixed five of a kind bonus counter.

* Add zeroed score indicator and end of game route update

 - Add indication using an asterisk to score category label for score
   that has been zeroed. Added attributes to scorepad object to track.
 - pytest validated zeroed score label process.
 - updated game process to proceed directly to end of game endpoint
   after last available score category has been selected.

* Update score display

Format and add borders to score sections. Further changes pending.

* Update menu item label for Total All Dice

Update category selection label for totalling all dice in a final roll.
Also update label for related score category in the score display.

* Fix free roll issue with five of a kind bonus category

When player selects a category, the five of a kind bonus
category allows player to select it regardless of player
having a five of a kind score and effectively finish the
turn without scoring or zeroing a score if there is at least
one score category available. This update checks to see if
a five of a kind has been scored and that it is not zero
before appending it to the available score category menu
list.

* Add score display to roll_one.html page

This will include the current score information for the
player to evaluate during a turn rolling dice.

* Update score display

Re-build score dispaly for scoring categories utilizing Bootstrap
flex grid attributes. Score display is included in the roll_one and the
score_display_and_select HTML pages.

* Fix Five of a Kind Bonus Selection

Add condition so that a menu choice for the five of a kind
bonus will only appear if the dice being evaluated is a five of
a kind roll.

* Refactor scorepad object in webgame.py

Update date any instance of scorepad to scorepad_global. This
should utilize the global object in any module.

* Create function to reset scorepad

This will allow a player to start a new game by restting the
attributes for the scorepad object.

* Make changes to href links and labels

Fix exit button in updated score html to properly go to
exit game endpoint. Some label renaming.

* Update index.html page and clean up default templates folder

Index.html cleaned up and the web/templates folder has been cleared of
any unused HTML files.

* Update Dice Selection HTML

Change UI to allow player to select dice using the labels on top
of the dice png's.

In backend, switch from using a WTForm to a standard POST request. Using
the collected request form data, add the kept dice and add to new roll.

Reorganize button placement on dice selection HTML.