Update Three of kind and Four of a kind scoring (Issue #4 fix)
Allow five of a kind dice to be scored as three or four of a kind.
This would apply to a case where the player zeroed their five of a
kind score, but end up with five of a kind dice later in the game.
With the fix the player can apply the five of a kind dice to either
an available three or four of a kind score.